+++ examples/fitness/pubspec.yaml
@@ -2,7 +2,7 @@ name: fitness
 dependencies:
   sky: any
   sky_tools: any
-  playfair: ^0.0.7
+  playfair: ^0.0.8
   path: ^1.3.6
 dependency_overrides:
   material_design_icons:
