+++ examples/game/test_drawatlas.dart
@@ -1,8 +1,8 @@
 import 'dart:sky';
 
-import 'package:sky/services.dart';
+import 'package:sky/material.dart';
 import 'package:sky/rendering.dart';
-import 'package:sky/theme/colors.dart' as colors;
+import 'package:sky/services.dart';
 import 'package:sky/widgets.dart';
 import 'package:skysprites/skysprites.dart';
 
@@ -37,7 +37,7 @@ class TestDrawAtlasApp extends App {
   Widget build() {
     ThemeData theme = new ThemeData(
       brightness: ThemeBrightness.light,
-      primarySwatch: colors.Purple
+      primarySwatch: Colors.purple
     );
 
     return new Theme(
