+++ examples/game/test_performance.dart
@@ -1,9 +1,9 @@
 import 'dart:sky' as sky;
 import 'dart:math' as math;
 
-import 'package:sky/services.dart';
+import 'package:sky/material.dart';
 import 'package:sky/rendering.dart';
-import 'package:sky/theme/colors.dart' as colors;
+import 'package:sky/services.dart';
 import 'package:sky/widgets.dart';
 import 'package:skysprites/skysprites.dart';
 
@@ -38,7 +38,7 @@ class TestApp extends App {
   Widget build() {
     ThemeData theme = new ThemeData(
       brightness: ThemeBrightness.light,
-      primarySwatch: colors.Purple
+      primarySwatch: Colors.purple
     );
 
     return new Theme(
