+++ examples/widgets/ensure_visible.dart
@@ -2,7 +2,7 @@
 // Use of this source code is governed by a BSD-style license that can be
 // found in the LICENSE file.
 
-import 'package:sky/theme/colors.dart' as colors;
+import 'package:sky/material.dart';
 import 'package:sky/widgets.dart';
 
 class CardModel {
@@ -17,10 +17,10 @@ class CardModel {
 class EnsureVisibleApp extends App {
 
   static const TextStyle cardLabelStyle =
-    const TextStyle(color: colors.white, fontSize: 18.0, fontWeight: bold);
+    const TextStyle(color: Colors.white, fontSize: 18.0, fontWeight: bold);
 
   static const TextStyle selectedCardLabelStyle =
-    const TextStyle(color: colors.white, fontSize: 24.0, fontWeight: bold);
+    const TextStyle(color: Colors.white, fontSize: 24.0, fontWeight: bold);
 
   List<CardModel> cardModels;
   MixedViewportLayoutState layoutState = new MixedViewportLayoutState();
@@ -33,7 +33,7 @@ class EnsureVisibleApp extends App {
       48.0, 63.0, 82.0, 146.0, 60.0, 55.0, 84.0, 96.0, 50.0
     ];
     cardModels = new List.generate(cardHeights.length, (i) {
-      Color color = Color.lerp(colors.Red[300], colors.Blue[900], i / cardHeights.length);
+      Color color = Color.lerp(Colors.red[300], Colors.blue[900], i / cardHeights.length);
       return new CardModel(i, cardHeights[i], color);
     });
 
@@ -83,8 +83,8 @@ class EnsureVisibleApp extends App {
       child: new Theme(
         data: new ThemeData(
           brightness: ThemeBrightness.light,
-          primarySwatch: colors.Blue,
-          accentColor: colors.RedAccent[200]
+          primarySwatch: Colors.blue,
+          accentColor: Colors.redAccent[200]
         ),
         child: new Title(
           title: 'Cards',
