+++ sky/packages/sky/lib/material.dart
@@ -1,0 +1,15 @@
+// Copyright 2015 The Chromium Authors. All rights reserved.
+// Use of this source code is governed by a BSD-style license that can be
+// found in the LICENSE file.
+
+/// Widgets implementing Material Design
+///
+/// See https://www.google.com/design/spec/material-design/introduction.html
+library material;
+
+export 'src/material/colors.dart';
+export 'src/material/edges.dart';
+export 'src/material/shadows.dart';
+export 'src/material/theme_data.dart';
+export 'src/material/typography.dart';
+export 'src/material/constants.dart';
