+++ sky/packages/sky/lib/src/material/README.md
@@ -2,5 +2,5 @@ This directory provides constants for Material Design.
 
    http://www.google.com/design/spec/material-design/
 
-Files in this directory only depend on [painting.dart](../painting.dart) and
+Files in this directory only depend on [painting.dart](../../painting.dart) and
 `sky:dart`.
