+++ sky/unit/test/widget/box_decoration_test.dart
@@ -1,4 +1,4 @@
-import 'package:sky/theme/colors.dart' as colors;
+import 'package:sky/material.dart';
 import 'package:sky/widgets.dart';
 import 'package:test/test.dart';
 
@@ -14,7 +14,7 @@ void main() {
         decoration: new BoxDecoration(
           shape: Shape.circle,
           border: new Border.all(width: 10.0, color: const Color(0x80FF00FF)),
-          backgroundColor: colors.Teal[600]
+          backgroundColor: Colors.teal[600]
         )
       );
     });
